#57 - DeckGL Layer as entity
