+++ src/app/basemap/basemap.component.ts
@@ -18,13 +18,6 @@ import { BrowserModule } from "@angular/platform-browser";
 import { FormsModule } from "@angular/forms";
 import { AppComponent } from "../app.component";
 
-// **************************************
-//! DeckGL imports
-import { MapboxLayer } from "@deck.gl/mapbox";
-import { TripsLayer } from "@deck.gl/geo-layers";
-import * as mob from "../../assets/modules_json_backup/mobility_service.json";
-
-// **************************************
 
 @NgModule({
   imports: [BrowserModule, FormsModule],
@@ -46,8 +39,6 @@ export class BasemapComponent implements OnInit, AfterViewInit {
   mapKeyVisible: boolean;
   layers: CsLayer[] = [];
   intervalMap = {};
-
-  json: any = mob;
 
   // Map config
   center: any;
@@ -125,8 +116,6 @@ export class BasemapComponent implements OnInit, AfterViewInit {
     this.map.boxZoom.disable();
 
     this.map.on("load", event => {
-      this.deckgl();
-
       this.mapCanvas = this.map.getCanvasContainer();
       this.updateMapLayers(event);
     });
@@ -541,63 +530,5 @@ export class BasemapComponent implements OnInit, AfterViewInit {
       this.router.navigate([""]);
       this.authenticationService.logout();
     });
-  }
-
-  // **************************************
-  // https://github.com/uber/deck.gl/blob/master/docs/api-reference/mapbox/mapbox-layer.md
-  // https://github.com/uber/deck.gl/blob/master/docs/api-reference/mapbox/overview.md?source=post_page---------------------------#using-with-pure-js
-
-  private deckgl() {
-    const DATA_URL = {
-      TRIPS:
-        // "https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/trips/trips.json" // eslint-disable-line
-        "https://cityio.media.mit.edu/api/table/grasbrook/trips"
-      // "https://cityio.media.mit.edu/api/table/grasbrook/cityIO_Gama_Hamburg"
-    };
-
-    function renderLayers() {
-      let loopLength = 4 * 60 * 60;
-      let animationSpeed = 50;
-      const timestamp = Date.now() / 1000;
-      const loopTime = loopLength / animationSpeed;
-      let time = 30000 + ((timestamp % loopTime) / loopTime) * loopLength;
-
-      tripLayer.setProps({
-        currentTime: time
-      });
-    }
-
-    const tripLayer = new MapboxLayer({
-      type: TripsLayer,
-      id: "trips",
-      data: DATA_URL.TRIPS,
-      getPath: d => d.segments,
-      getColor: d => {
-        switch (d.mode) {
-          case 0:
-            return [255, 0, 255];
-          case 1:
-            return [60, 128, 255];
-          case 2:
-            return [153, 255, 51];
-          case 3:
-            return [153, 180, 100];
-        }
-      },
-      opacity: 0.5,
-      widthMinPixels: 4,
-      rounded: true,
-      trailLength: 500,
-      currentTime: 0
-    });
-
-    this.map.addLayer(tripLayer);
-
-    // start animation loop
-    setInterval(() => {
-      renderLayers();
-    });
-
-    // **************************************
   }
 }
