+++ src/app/services/layer-loader.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { CsLayer } from "../../typings";
 import { ConfigurationService } from "./configuration.service";
+import {TripsDeckGlLayer} from "../layers/trips.deck-gl.layer";
 
 @Injectable({
   providedIn: "root"
@@ -11,6 +12,13 @@ export class LayerLoaderService {
   getLayers(): CsLayer[] {
     const layers: CsLayer[] = [];
     Array.prototype.push.apply(layers, this.config.layers);
+
+    for (let layer of layers) {
+      if (layer.id === "trips-deckgl") {
+        let tripsDeck = TripsDeckGlLayer.createTripsLayer();
+        layers[layers.indexOf(layer)] = tripsDeck;
+      }
+    }
 
     return layers;
   }
